# Optimise text.lines for longer lines, larger chunks use-cases